+++ CHANGELOG.md
@@ -22,6 +22,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Add the
   `github.com/signalfx/splunk-otel-go/instrumentation/github.com/lib/pq/splunkpq`
   instrumentation for the `github.com/lib/pq` package. (#92)
+- Add the
+  `github.com/signalfx/splunk-otel-go/instrumentation/github.com/jmoiron/sqlx/splunksqlx`
+  instrumentation for the `github.com/jmoiron/sqlx` package. (#93)
 
 ### Changed